requirements-gui.txt
arithmos-canvas-core>=0.1.9,<0.2a
arithmos-widget-base>=4.2.0a

# PyQt4/PyQt5 compatibility
AnyQt>=0.0.8

pyqtgraph==0.10.0
matplotlib>=2.0.0

# For add-ons' descriptions
docutils
